componentized text box
